️➕ADD : New image and loggin google
