+++ .eslintrc.json
@@ -1,3 +1,8 @@
 {
-  "extends": "next/core-web-vitals"
+  "extends": "next/core-web-vitals",
+  "rules": {
+    "semi": "off",
+    "quotes": "off",
+    "react/no-unescaped-entities": "off"
+  }
 }
