--- src/api/auth/[...nextauth]/route.ts
@@ -1,13 +1,0 @@
-import NextAuth from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-
-const handler = NextAuth({
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID as string, //*Client ID es basicamente una especie de identificador para nuestra aplicacion para saver como conectarnos.
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
-    }),
-  ],
-});
-
-export { handler as GET, handler as POST };
