+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,20 @@
+import NextAuth from "next-auth/next";
+import GoogleProvider from "next-auth/providers/google";
+import FacebookProvider from "next-auth/providers/facebook";
+
+const authOptions = {
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID as string,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
+    }),
+    FacebookProvider({
+      clientId: process.env.FACEBOOK_CLIENT_ID as string,
+      clientSecret: process.env.FACEBOOK_CLIENT_SECRET as string,
+    }),
+  ],
+};
+
+const handler = NextAuth(authOptions)
+
+export {handler as GET, handler as POST}
