+++ src/app/layout.tsx
@@ -1,26 +1,31 @@
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import './globals.css'
-import Navbar from '@/components/Su-Gou/navbar'
+import Navbar from "@/components/Navbar";
+import {Providers} from "./providers";
 
-const inter = Inter({ subsets: ['latin'] })
+const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
     <html lang="en">
       <body className={inter.className}>
-      <Navbar/>
-        {children}
-        </body>
+        <Providers>
+         <div>
+         <Navbar />
+          {children}
+         </div>
+        </Providers>
+      </body>
     </html>
-  )
+  );
 }
