+++ src/app/providers.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import { SessionProvider } from "next-auth/react";
+
+export const Providers = ({ children }: { children: React.ReactNode }) => {
+  return <SessionProvider>{children}</SessionProvider>;
+}
